Added Model  auto detection and manual selection
