+++ Camera_Setup_Dump.txt
@@ -1,0 +1,199 @@
+HTTP Recived from PTZ: p1
+HTTP Recived from PTZ: OID:AW-UE150
+HTTP Recived from PTZ: CGI_TIME:0
+HTTP Recived from PTZ: OSA:87:0x11
+HTTP Recived from PTZ: TITLE:AW-UE150-2
+HTTP Recived from PTZ: OSJ:5C:41572D55453135302D3200000000000000000000
+HTTP Recived from PTZ: OSF:0
+HTTP Recived from PTZ: OSD:48:0x23
+HTTP Recived from PTZ: ORS:1
+HTTP Recived from PTZ: d31
+HTTP Recived from PTZ: OSJ:01:0x1
+HTTP Recived from PTZ: OSJ:02:0x0
+HTTP Recived from PTZ: OSJ:C0:0x0
+HTTP Recived from PTZ: OSJ:03:0x0
+HTTP Recived from PTZ: OSJ:06:0x003C
+HTTP Recived from PTZ: OSJ:09:0x001F4
+HTTP Recived from PTZ: OSD:BF:2
+HTTP Recived from PTZ: OGU:0x0F
+HTTP Recived from PTZ: OSI:28:0
+HTTP Recived from PTZ: OSD:69:03
+HTTP Recived from PTZ: OSA:65:0x00
+HTTP Recived from PTZ: OFT:0
+HTTP Recived from PTZ: d60
+HTTP Recived from PTZ: OAW:9
+HTTP Recived from PTZ: OSI:20:0x00A6E:0
+HTTP Recived from PTZ: OSG:39:0x800
+HTTP Recived from PTZ: OSG:3A:0x800
+HTTP Recived from PTZ: OSJ:4A:0x00C80:0
+HTTP Recived from PTZ: OSJ:4B:0x800
+HTTP Recived from PTZ: OSJ:4C:0x800
+HTTP Recived from PTZ: OSJ:4D:0x800
+HTTP Recived from PTZ: OSJ:0C:0x0
+HTTP Recived from PTZ: OSI:25:0
+HTTP Recived from PTZ: OSJ:0D:0x80
+HTTP Recived from PTZ: OSJ:0E:0x80
+HTTP Recived from PTZ: OSD:B0:0x80
+HTTP Recived from PTZ: OSJ:0B:0x80
+HTTP Recived from PTZ: OSJ:0F:0x800
+HTTP Recived from PTZ: ORP:0x09A
+HTTP Recived from PTZ: OSJ:10:0x096
+HTTP Recived from PTZ: OBP:0x096
+HTTP Recived from PTZ: OSJ:11:0
+HTTP Recived from PTZ: ODT:0
+HTTP Recived from PTZ: OSA:30:80
+HTTP Recived from PTZ: OSJ:12:0x0F
+HTTP Recived from PTZ: OSD:A1:0x80
+HTTP Recived from PTZ: OSD:A2:0x80
+HTTP Recived from PTZ: OSJ:13:0x80
+HTTP Recived from PTZ: OSG:3F:0x02
+HTTP Recived from PTZ: OSA:38:0x80
+HTTP Recived from PTZ: OSA:39:0x80
+HTTP Recived from PTZ: OSA:40:0
+HTTP Recived from PTZ: OSD:A3:0x90
+HTTP Recived from PTZ: OSJ:14:0x1
+HTTP Recived from PTZ: OSJ:15:0x6C
+HTTP Recived from PTZ: OSJ:16:0x0F
+HTTP Recived from PTZ: OSJ:17:0x87
+HTTP Recived from PTZ: OSJ:18:0x80
+HTTP Recived from PTZ: OSJ:19:0x80
+HTTP Recived from PTZ: OSJ:1A:0x02
+HTTP Recived from PTZ: OSE:72:0
+HTTP Recived from PTZ: OSA:6A:0x76
+HTTP Recived from PTZ: OSA:10:3
+HTTP Recived from PTZ: OSA:0F:0x00
+HTTP Recived from PTZ: OSA:25:0x7C
+HTTP Recived from PTZ: OSA:21:0x62
+HTTP Recived from PTZ: OSA:07:0x80
+HTTP Recived from PTZ: OSJ:1B:0x1
+HTTP Recived from PTZ: OSE:33:0
+HTTP Recived from PTZ: OSA:2D:2
+HTTP Recived from PTZ: OSG:97:4
+HTTP Recived from PTZ: OSA:20:0x7E
+HTTP Recived from PTZ: OSA:24:0x63
+HTTP Recived from PTZ: OSI:40:0
+HTTP Recived from PTZ: OSI:41:0x1C
+HTTP Recived from PTZ: OSI:42:0x0A
+HTTP Recived from PTZ: OSA:2E:1
+HTTP Recived from PTZ: OSA:2A:0x13
+HTTP Recived from PTZ: OSD:3A:0x01
+HTTP Recived from PTZ: OSE:31:0
+HTTP Recived from PTZ: OSJ:4F:0
+HTTP Recived from PTZ: OSD:A4:0x80
+HTTP Recived from PTZ: OSD:A5:0x80
+HTTP Recived from PTZ: OSD:A6:0x90
+HTTP Recived from PTZ: OSD:A7:0x93
+HTTP Recived from PTZ: OSD:A8:0x80
+HTTP Recived from PTZ: OSD:A9:0x92
+HTTP Recived from PTZ: OSD:80:0x8F
+HTTP Recived from PTZ: OSD:81:0x86
+HTTP Recived from PTZ: OSD:82:0x8D
+HTTP Recived from PTZ: OSD:83:0x8A
+HTTP Recived from PTZ: OSD:84:0x84
+HTTP Recived from PTZ: OSD:85:0x8A
+HTTP Recived from PTZ: OSD:9A:0x82
+HTTP Recived from PTZ: OSD:9B:0x86
+HTTP Recived from PTZ: OSD:86:0x80
+HTTP Recived from PTZ: OSD:87:0x81
+HTTP Recived from PTZ: OSD:9C:0x84
+HTTP Recived from PTZ: OSD:9D:0x86
+HTTP Recived from PTZ: OSD:88:0x85
+HTTP Recived from PTZ: OSD:89:0x8B
+HTTP Recived from PTZ: OSD:9E:0x8A
+HTTP Recived from PTZ: OSD:9F:0x83
+HTTP Recived from PTZ: OSD:8A:0x88
+HTTP Recived from PTZ: OSD:8B:0x80
+HTTP Recived from PTZ: OSJ:1C:0x83
+HTTP Recived from PTZ: OSJ:1D:0x82
+HTTP Recived from PTZ: OSD:8C:0x92
+HTTP Recived from PTZ: OSD:8D:0x88
+HTTP Recived from PTZ: OSD:8E:0x94
+HTTP Recived from PTZ: OSD:8F:0x90
+HTTP Recived from PTZ: OSD:90:0x8F
+HTTP Recived from PTZ: OSD:91:0x92
+HTTP Recived from PTZ: OSD:92:0x89
+HTTP Recived from PTZ: OSD:93:0x88
+HTTP Recived from PTZ: OSD:94:0x81
+HTTP Recived from PTZ: OSD:95:0x8B
+HTTP Recived from PTZ: OSD:96:0x85
+HTTP Recived from PTZ: OSD:97:0x76
+HTTP Recived from PTZ: OAF:0
+HTTP Recived from PTZ: d10
+HTTP Recived from PTZ: OSD:B3:0
+HTTP Recived from PTZ: OSE:70:0
+HTTP Recived from PTZ: OSE:7A:10
+HTTP Recived from PTZ: ODE:0
+HTTP Recived from PTZ: OSJ:4E:0
+HTTP Recived from PTZ: OIS:0
+HTTP Recived from PTZ: axz557
+HTTP Recived from PTZ: axf64C
+HTTP Recived from PTZ: axiFE9
+HTTP Recived from PTZ: OSD:4F:FC
+HTTP Recived from PTZ: OIF:0x1C
+HTTP Recived from PTZ: OSE:77:1
+HTTP Recived from PTZ: fAN0
+HTTP Recived from PTZ: fA20
+HTTP Recived from PTZ: OBR:0
+HTTP Recived from PTZ: tAE1
+HTTP Recived from PTZ: TLR:0
+HTTP Recived from PTZ: dA0
+HTTP Recived from PTZ: TLG:0
+HTTP Recived from PTZ: tAA000000000
+HTTP Recived from PTZ: OSJ:2E:0x0
+HTTP Recived from PTZ: OSI:32:0
+HTTP Recived from PTZ: OSI:33:1
+HTTP Recived from PTZ: OSI:16:1
+HTTP Recived from PTZ: OSI:1A:7
+HTTP Recived from PTZ: OSI:17:1
+HTTP Recived from PTZ: OSJ:2F:0x3C0
+HTTP Recived from PTZ: OSJ:30:0x21C
+HTTP Recived from PTZ: OSJ:31:0x3C0
+HTTP Recived from PTZ: OSJ:32:0x21C
+HTTP Recived from PTZ: OSJ:33:0x3C0
+HTTP Recived from PTZ: OSJ:34:0x21C
+HTTP Recived from PTZ: iNS0
+HTTP Recived from PTZ: OSJ:2D:0
+HTTP Recived from PTZ: sWZ1
+HTTP Recived from PTZ: OAZ:0
+HTTP Recived from PTZ: lC10
+HTTP Recived from PTZ: lC20
+HTTP Recived from PTZ: lC30
+HTTP Recived from PTZ: lC40
+HTTP Recived from PTZ: OSJ:29:0
+HTTP Recived from PTZ: pST2
+HTTP Recived from PTZ: uPVS999
+HTTP Recived from PTZ: OSJ:A8:1
+HTTP Recived from PTZ: OSJ:AB:0x03
+HTTP Recived from PTZ: OSE:71:0
+HTTP Recived from PTZ: OSE:7C:0
+HTTP Recived from PTZ: OSJ:2A:0x0
+HTTP Recived from PTZ: OSJ:2B:0x1
+HTTP Recived from PTZ: OSJ:2C:0x0
+HTTP Recived from PTZ: OSJ:5B:0
+HTTP Recived from PTZ: OSE:7D:0
+HTTP Recived from PTZ: pRF0
+HTTP Recived from PTZ: pE000000000313
+HTTP Recived from PTZ: pE010000000000
+HTTP Recived from PTZ: pE020000000000
+HTTP Recived from PTZ: s0
+HTTP Recived from PTZ: OUS:0
+HTTP Recived from PTZ: OSJ:40:0:
+HTTP Recived from PTZ: OER:0
+HTTP Recived from PTZ: OSI:46:0x00000000
+HTTP Recived from PTZ: rER00
+HTTP Recived from PTZ: aPC80007C00
+HTTP Recived from PTZ: OSJ:45:2
+HTTP Recived from PTZ: OSJ:46:00
+HTTP Recived from PTZ: OSI:30:1
+HTTP Recived from PTZ: OSJ:54:0
+HTTP Recived from PTZ: OSJ:55:0
+HTTP Recived from PTZ: OSJ:C1:0
+HTTP Recived from PTZ: OSJ:56:0
+HTTP Recived from PTZ: OSJ:1F:0
+HTTP Recived from PTZ: OSJ:57:0
+HTTP Recived from PTZ: OSJ:22:0
+HTTP Recived from PTZ: OSJ:58:0
+HTTP Recived from PTZ: OSJ:24:0
+HTTP Recived from PTZ: OSJ:59:0
+HTTP Recived from PTZ: OSJ:26:0
+HTTP Recived from PTZ: OSJ:5A:0
