+++ package.json
@@ -1,30 +1,18 @@
 {
 	"name": "panasonic-ptz",
 	"version": "1.0.9",
-	"api_version": "1.0.0",
+	"api_version": "1.12",
 	"keywords": [
 		"ptz",
 		"Camera"
 	],
 	"manufacturer": "Panasonic",
 	"product": [
-		"AW-HE120",
-		"AW-HE130",
-		"AW-HE40",
-		"AW-HE42",
-		"AW-HE50",
-		"AW-HE60",
-		"AW-HE65",
-		"AW-HE70",
-		"AW-UE70",
-		"AW-UE150",
-		"AW-HR140",
-		"AW-SFU01",
-		"AK-UB300"
+		"PTZ Camera"
 	],
 	"shortname": "Panasonic-PTZ",
 	"description": "Companion module for Panasonic PTZ Cameras",
-	"main": "index.js",
+	"main": "src/index.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1"
 	},
